requirements.txt
Flask==3.1.3
flask-cors==6.0.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
web3==7.14.1
